Move task to a crate.
