+++ crates/action-runner/Cargo.toml
@@ -11,6 +11,7 @@ moon_lang = { path = "../lang" }
 moon_logger = { path = "../logger" }
 moon_project = { path = "../project" }
 moon_project_graph = { path = "../project-graph" }
+moon_task = { path = "../task" }
 moon_terminal = { path = "../terminal" }
 moon_utils = { path = "../utils" }
 moon_workspace = { path = "../workspace" }
