+++ crates/action/src/errors.rs
@@ -1,5 +1,6 @@
 use moon_error::MoonError;
 use moon_project::ProjectError;
+use moon_task::{TargetError, TaskError};
 use moon_toolchain::ToolchainError;
 use moon_vcs::VcsError;
 use moon_workspace::WorkspaceError;
@@ -12,6 +13,12 @@ pub enum ActionError {
 
     #[error(transparent)]
     Project(#[from] ProjectError),
+
+    #[error(transparent)]
+    Task(#[from] TaskError),
+
+    #[error(transparent)]
+    Target(#[from] TargetError),
 
     #[error(transparent)]
     Toolchain(#[from] ToolchainError),
