+++ crates/hasher/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 moon_config = { path = "../config"}
 moon_lang_node = { path = "../lang-node" }
 moon_project = { path = "../project"}
+moon_task = { path = "../task"}
 moon_utils = { path = "../utils"}
 serde = { version = "1.0.138", features = ["derive"] }
 sha2 = "0.10.2"
