+++ crates/project/src/lib.rs
@@ -1,28 +1,9 @@
 mod errors;
-mod file_group;
 mod helpers;
 mod project;
-mod target;
-mod task;
-pub mod test;
-mod token;
-mod types;
 
-pub use errors::{ProjectError, TargetError};
+pub use errors::ProjectError;
 pub use helpers::*;
-pub use types::*;
 
-// Projects
 pub use moon_config::{ProjectID, ProjectType};
 pub use project::{Project, ProjectsMap};
-
-// Tasks & targets
-pub use moon_config::{TargetID, TaskID, TaskType};
-pub use target::{Target, TargetProject};
-pub use task::{Task, TaskOptions};
-
-// Tokens
-pub use token::{ResolverType, TokenResolver, TokenSharedData, TokenType};
-
-// File groups
-pub use file_group::FileGroup;
