+++ crates/task/src/lib.rs
@@ -1,0 +1,16 @@
+mod errors;
+mod file_group;
+mod target;
+mod task;
+pub mod test;
+mod token;
+mod types;
+
+pub use moon_config::{TargetID, TaskID, TaskType};
+
+pub use errors::*;
+pub use file_group::FileGroup;
+pub use target::{Target, TargetProjectScope};
+pub use task::{Task, TaskOptions};
+pub use token::{ResolverType, TokenResolver, TokenSharedData, TokenType};
+pub use types::*;
