+++ crates/task/src/task.rs
@@ -1,5 +1,5 @@
-use crate::errors::{ProjectError, TargetError};
-use crate::target::{Target, TargetProject};
+use crate::errors::{TargetError, TaskError};
+use crate::target::{Target, TargetProjectScope};
 use crate::token::TokenResolver;
 use crate::types::{EnvVars, ExpandedFiles, TouchedFilePaths};
 use moon_config::{
@@ -148,12 +148,12 @@ impl Task {
     }
 
     /// Create a globset of all input globs to match with.
-    pub fn create_globset(&self) -> Result<glob::GlobSet, ProjectError> {
+    pub fn create_globset(&self) -> Result<glob::GlobSet, TaskError> {
         Ok(glob::GlobSet::new(&self.input_globs)?)
     }
 
     /// Expand the args list to resolve tokens, relative to the project root.
-    pub fn expand_args(&mut self, token_resolver: TokenResolver) -> Result<(), ProjectError> {
+    pub fn expand_args(&mut self, token_resolver: TokenResolver) -> Result<(), TaskError> {
         if self.args.is_empty() {
             return Ok(());
         }
@@ -204,11 +204,7 @@ impl Task {
     }
 
     /// Expand the deps list and resolve parent/self scopes.
-    pub fn expand_deps(
-        &mut self,
-        owner_id: &str,
-        depends_on: &[String],
-    ) -> Result<(), ProjectError> {
+    pub fn expand_deps(&mut self, owner_id: &str, depends_on: &[String]) -> Result<(), TaskError> {
         if self.deps.is_empty() {
             return Ok(());
         }
@@ -227,17 +223,17 @@ impl Task {
 
             match &target.project {
                 // ^:task
-                TargetProject::Deps => {
+                TargetProjectScope::Deps => {
                     for project_id in depends_on {
                         push_dep(Target::format(project_id, &target.task_id)?);
                     }
                 }
                 // ~:task
-                TargetProject::Own => {
+                TargetProjectScope::Own => {
                     push_dep(Target::format(owner_id, &target.task_id)?);
                 }
                 // project:task
-                TargetProject::Id(_) => {
+                TargetProjectScope::Id(_) => {
                     push_dep(dep.clone());
                 }
                 _ => {
@@ -252,7 +248,7 @@ impl Task {
     }
 
     /// Expand the inputs list to a set of absolute file paths, while resolving tokens.
-    pub fn expand_inputs(&mut self, token_resolver: TokenResolver) -> Result<(), ProjectError> {
+    pub fn expand_inputs(&mut self, token_resolver: TokenResolver) -> Result<(), TaskError> {
         if self.inputs.is_empty() {
             return Ok(());
         }
@@ -270,14 +266,14 @@ impl Task {
     }
 
     /// Expand the outputs list to a set of absolute file paths, while resolving tokens.
-    pub fn expand_outputs(&mut self, token_resolver: TokenResolver) -> Result<(), ProjectError> {
+    pub fn expand_outputs(&mut self, token_resolver: TokenResolver) -> Result<(), TaskError> {
         if self.outputs.is_empty() {
             return Ok(());
         }
 
         for output in &token_resolver.resolve(&self.outputs, self)? {
             if glob::is_path_glob(output) {
-                return Err(ProjectError::NoOutputGlob(
+                return Err(TaskError::NoOutputGlob(
                     output.to_owned(),
                     self.target.clone(),
                 ));
@@ -291,7 +287,7 @@ impl Task {
 
     /// Return true if this task is affected, based on touched files.
     /// Will attempt to find any file that matches our list of inputs.
-    pub fn is_affected(&self, touched_files: &TouchedFilePaths) -> Result<bool, ProjectError> {
+    pub fn is_affected(&self, touched_files: &TouchedFilePaths) -> Result<bool, TaskError> {
         let has_globs = !self.input_globs.is_empty();
         let globset = self.create_globset()?;
 
@@ -413,135 +409,5 @@ impl Task {
         }
 
         list
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::test::create_expanded_task;
-    use moon_config::TaskConfig;
-    use moon_utils::string_vec;
-    use moon_utils::test::get_fixtures_dir;
-    use std::collections::HashSet;
-
-    #[test]
-    #[should_panic(expected = "NoOutputGlob")]
-    fn errors_for_output_glob() {
-        let workspace_root = get_fixtures_dir("projects");
-        let project_root = workspace_root.join("basic");
-
-        create_expanded_task(
-            &workspace_root,
-            &project_root,
-            Some(TaskConfig {
-                outputs: Some(string_vec!["some/**/glob"]),
-                ..TaskConfig::default()
-            }),
-        )
-        .unwrap();
-    }
-
-    mod is_affected {
-        use super::*;
-
-        #[test]
-        fn returns_true_if_matches_file() {
-            let workspace_root = get_fixtures_dir("base");
-            let project_root = workspace_root.join("files-and-dirs");
-            let task = create_expanded_task(
-                &workspace_root,
-                &project_root,
-                Some(TaskConfig {
-                    inputs: Some(string_vec!["file.ts"]),
-                    ..TaskConfig::default()
-                }),
-            )
-            .unwrap();
-
-            let mut set = HashSet::new();
-            set.insert(project_root.join("file.ts"));
-
-            assert!(task.is_affected(&set).unwrap());
-        }
-
-        #[test]
-        fn returns_true_if_matches_glob() {
-            let workspace_root = get_fixtures_dir("base");
-            let project_root = workspace_root.join("files-and-dirs");
-            let task = create_expanded_task(
-                &workspace_root,
-                &project_root,
-                Some(TaskConfig {
-                    inputs: Some(string_vec!["file.*"]),
-                    ..TaskConfig::default()
-                }),
-            )
-            .unwrap();
-
-            let mut set = HashSet::new();
-            set.insert(project_root.join("file.ts"));
-
-            assert!(task.is_affected(&set).unwrap());
-        }
-
-        #[test]
-        fn returns_true_when_referencing_root_files() {
-            let workspace_root = get_fixtures_dir("base");
-            let project_root = workspace_root.join("files-and-dirs");
-            let task = create_expanded_task(
-                &workspace_root,
-                &project_root,
-                Some(TaskConfig {
-                    inputs: Some(string_vec!["/package.json"]),
-                    ..TaskConfig::default()
-                }),
-            )
-            .unwrap();
-
-            let mut set = HashSet::new();
-            set.insert(workspace_root.join("package.json"));
-
-            assert!(task.is_affected(&set).unwrap());
-        }
-
-        #[test]
-        fn returns_false_if_outside_project() {
-            let workspace_root = get_fixtures_dir("base");
-            let project_root = workspace_root.join("files-and-dirs");
-            let task = create_expanded_task(
-                &workspace_root,
-                &project_root,
-                Some(TaskConfig {
-                    inputs: Some(string_vec!["file.ts"]),
-                    ..TaskConfig::default()
-                }),
-            )
-            .unwrap();
-
-            let mut set = HashSet::new();
-            set.insert(workspace_root.join("base/other/outside.ts"));
-
-            assert!(!task.is_affected(&set).unwrap());
-        }
-
-        #[test]
-        fn returns_false_if_no_match() {
-            let workspace_root = get_fixtures_dir("base");
-            let project_root = workspace_root.join("files-and-dirs");
-            let task = create_expanded_task(
-                &workspace_root,
-                &project_root,
-                Some(TaskConfig {
-                    inputs: Some(string_vec!["file.ts", "src/*"]),
-                    ..TaskConfig::default()
-                }),
-            )
-            .unwrap();
-
-            let mut set = HashSet::new();
-            set.insert(project_root.join("another.rs"));
-
-            assert!(!task.is_affected(&set).unwrap());
-        }
     }
 }
