+++ crates/task/src/types.rs
@@ -1,4 +1,3 @@
-use moon_config::ProjectID;
 use std::collections::{HashMap, HashSet};
 use std::path::PathBuf;
 
@@ -7,5 +6,3 @@ pub type TouchedFilePaths = HashSet<PathBuf>;
 pub type ExpandedFiles = HashSet<PathBuf>;
 
 pub type EnvVars = HashMap<String, String>;
-
-pub type ProjectsSourceMap = HashMap<ProjectID, String>;
